1.sql파일 추가( Table 생성에 대한 쿼리문 있음) 2.Restaurant, imageboard, ReviewBoard 수정  --> Account  엔티티 클래스 수정 필요.(현규형!)

+++ src/main/java/com/aloa/review/ImageRepository.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<ImageBoard,Long>{
-	List<ImageBoard> findByReviewNoLessThanOrderByCreateDateDesc(Long reviewNo);
+public interface ImageRepository extends JpaRepository<ImageBoard,Integer>{
+	List<ImageBoard> findByReviewNoLessThanOrderByCreateDateDesc(int reviewNo);
 }
